This commit was manufactured by cvs2svn to create tag 'STLPORT_5_0_RC1'.
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/tags/STLPORT_5_0_RC1/STLport@759 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/regression/rotcopy1.cpp
@@ -25,7 +25,7 @@ int rotcopy1_test(int, char**)
   copy(v1.begin(), v1.end(), iter);
   cout << endl;
   vector <int> v2(v1.size());
-  for(int i = 0; i < v1.size(); i++)
+  for(size_t i = 0; i < v1.size(); i++)
   {
     rotate_copy(v1.begin(),
                  v1.begin() + i,
